# fake_quant: more memory efficient per-channel backward