--- a/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
+++ b/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
@@ -2151,6 +2151,35 @@ void fake_quant_per_channel_cpu(
   });
 }
 
+void fake_quant_per_channel_cachemask_cpu(
+    TensorIterator& iter,
+    TensorIterator& iter_mask,
+    int64_t quant_min,
+    int64_t quant_max) {
+  // TODO(future, optional): read once, write twice.  Not done at the moment
+  //   for simplicity, as we do not expect this to be a bottleneck.
+
+  // write mask
+  cpu_kernel(iter_mask, [=](float self, float scale, int64_t zero_point) -> bool {
+    float inv_scale = 1.0f / scale;
+    const auto qval = static_cast<int64_t>(zero_point + std::nearbyint(self * inv_scale));
+    return ((quant_min <= qval) && (qval <= quant_max));
+  });
+
+  // write fake_quant
+  cpu_kernel(iter, [=](float self, float scale, int64_t zero_point) -> float {
+    float inv_scale = 1.0f / scale;
+    return (std::fmin(
+                std::fmax(
+                    static_cast<int64_t>(
+                        zero_point + std::nearbyint(self * inv_scale)),
+                    quant_min),
+                quant_max) -
+            zero_point) *
+        scale;
+  });
+}
+
 void fake_quant_grad_per_channel_cpu(
     TensorIterator& iter,
     int64_t quant_min,
@@ -3046,6 +3075,7 @@ REGISTER_DISPATCH(fake_quant_grad_learnable_tensor_stub,
 REGISTER_DISPATCH(fake_quant_grad_per_channel_stub,
                   &fake_quant_grad_per_channel_cpu);
 REGISTER_DISPATCH(fake_quant_per_channel_stub, &fake_quant_per_channel_cpu);
+REGISTER_DISPATCH(fake_quant_per_channel_cachemask_stub, &fake_quant_per_channel_cachemask_cpu);
 REGISTER_DISPATCH(fake_quant_tensor_cachemask_stub,
                   &fake_quantize_tensor_cachemask_kernel);
 REGISTER_DISPATCH(qadaptive_avg_pool2d_nhwc_stub,