--- a/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
+++ b/aten/src/ATen/native/quantized/cuda/fake_quantize_core.cu
@@ -99,6 +99,35 @@ void fake_quant_per_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_
     });
 }
 
+void fake_quant_per_channel_cachemask_cuda(
+    TensorIterator &iter, TensorIterator &iter_mask, int64_t quant_min, int64_t quant_max) {
+  // TODO(future, optional): read once, write twice.  Not done at the moment
+  //   for simplicity, as we do not expect this to be a bottleneck.
+
+  // write mask
+  gpu_kernel(iter_mask,
+    [=] GPU_LAMBDA (float input_val, float scale, int64_t zero_point) -> bool {
+      float inv_scale = 1.0f / scale;
+      const auto qval = static_cast<int64_t>(std::nearbyint(input_val * inv_scale) + zero_point);
+      return ((quant_min <= qval) && (qval <= quant_max));
+    });
+
+  // write fake_quant
+  gpu_kernel(iter,
+    [=] GPU_LAMBDA (float input_val, float scale, int64_t zero_point) -> float {
+      float inv_scale = 1.0f / scale;
+      return (fminf(
+                  quant_max,
+                  fmaxf(
+                      quant_min,
+                      static_cast<int64_t>(
+                          std::nearbyint(input_val * inv_scale) +
+                          zero_point))) -
+              zero_point) *
+          scale;
+    });
+}
+
 void fake_quant_grad_per_channel_cuda(TensorIterator &iter, int64_t quant_min, int64_t quant_max) {
   gpu_kernel(iter,
     [=] GPU_LAMBDA (float x, float dy, float scale, int64_t zero_point) -> float {
@@ -134,6 +163,7 @@ void _fake_quantize_grad_learnable_channel_kernel_cuda(TensorIterator &iter, int
 
 REGISTER_DISPATCH(fake_quant_per_channel_stub, &fake_quant_per_channel_cuda);
 REGISTER_DISPATCH(fake_quant_grad_per_channel_stub, &fake_quant_grad_per_channel_cuda);
+REGISTER_DISPATCH(fake_quant_per_channel_cachemask_stub, &fake_quant_per_channel_cachemask_cuda);
 REGISTER_DISPATCH(fake_quant_grad_learnable_channel_stub, &_fake_quantize_grad_learnable_channel_kernel_cuda);
 
 } // namespace native